Extract shared elapsed time display
