+++ src/browser/features/Tools/BashToolCall.tsx
@@ -18,6 +18,7 @@ import {
 import { useToolExpansion, getStatusDisplay, type ToolStatus } from "./Shared/toolUtils";
 import { formatDuration } from "@/common/utils/formatDuration";
 import { cn } from "@/common/lib/utils";
+import { ElapsedTimeDisplay } from "./Shared/ElapsedTimeDisplay";
 import { useBashToolLiveOutput, useLatestStreamingBashId } from "@/browser/stores/WorkspaceStore";
 import { useForegroundBashToolCallIds } from "@/browser/stores/BackgroundBashStore";
 import { useBackgroundBashActions } from "@/browser/contexts/BackgroundBashContext";
@@ -32,64 +33,6 @@ interface BashToolCallProps {
   status?: ToolStatus;
   startedAt?: number;
 }
-
-/**
- * Isolated component for elapsed time display.
- * Uses requestAnimationFrame + local state to avoid re-rendering parent component.
- */
-const ElapsedTimeDisplay: React.FC<{ startedAt: number | undefined; isActive: boolean }> = ({
-  startedAt,
-  isActive,
-}) => {
-  const elapsedRef = useRef(0);
-  const frameRef = useRef<number | null>(null);
-  const [, forceUpdate] = React.useReducer((x: number) => x + 1, 0);
-  const baseStart = useRef(startedAt ?? Date.now());
-
-  useEffect(() => {
-    if (!isActive) {
-      elapsedRef.current = 0;
-      if (frameRef.current !== null) {
-        cancelAnimationFrame(frameRef.current);
-        frameRef.current = null;
-      }
-      return;
-    }
-
-    baseStart.current = startedAt ?? Date.now();
-    let lastSecond = -1;
-
-    const tick = () => {
-      const now = Date.now();
-      const elapsed = now - baseStart.current;
-      const currentSecond = Math.floor(elapsed / 1000);
-
-      // Only update when second changes to minimize renders
-      if (currentSecond !== lastSecond) {
-        lastSecond = currentSecond;
-        elapsedRef.current = elapsed;
-        forceUpdate();
-      }
-
-      frameRef.current = requestAnimationFrame(tick);
-    };
-
-    tick();
-
-    return () => {
-      if (frameRef.current !== null) {
-        cancelAnimationFrame(frameRef.current);
-        frameRef.current = null;
-      }
-    };
-  }, [isActive, startedAt]);
-
-  if (!isActive || elapsedRef.current === 0) {
-    return null;
-  }
-
-  return <> • {Math.round(elapsedRef.current / 1000)}s</>;
-};
 
 type BashLiveOutputView = NonNullable<ReturnType<typeof useBashToolLiveOutput>>;
